Remove deployment on staging on pull request
It is not necessary to deploy on pull request because the PR may or may not work (which we do not want to deploy). When the feature is working, we will merge the PR which will necessarily need a push (and this push will be deployed, which is what we want).